Adds bootstrap css and refactors home component
