+++ lib/filter-tile/description.js
@@ -7,9 +7,9 @@ export default class Description extends Component {
     const { description } = this.props;
     return (
       <div className='descriptionTextContainer'>
-        <div className='description'>{description.part1}</div>
-        <div className='bigWord'>{description.bigWord}</div>
-        <div className='description'>{description.part2}</div>
+        <div className='descriptionText'>{description.part1}</div>
+        <div className='bigCapsWord'>{description.bigWord}</div>
+        <div className='descriptionText'>{description.part2}</div>
       </div>
     );
   }
