+++ lib/package-tile/index.js
@@ -27,7 +27,7 @@ export default class PackageTile extends Component {
     });
 
     return (
-      <div className='container'>
+      <div className='packageContainer'>
         <div className='packageImage' style={{backgroundImage: `url(${imageSrc})`}}>
           <div className='mintBar'>{caption}</div>
         </div>
@@ -40,11 +40,11 @@ export default class PackageTile extends Component {
             <div className='ratingDenominator'>{'/' + rating.split('/')[1]}</div>
           </div>
           <div className='bottomContainer'>
-            <div className='row'>
+            <div className='descriptionRow'>
               <div className='packageDetailsText'>{`✔ ${packageDetails}`}</div>
               <div className='strikethroughText'>{oldPrice}</div>
             </div>
-            <div className='row'>
+            <div className='descriptionRow'>
               <div className='date'>{period}</div>
               <div className='price'>{`${newPrice}:-`}</div>
             </div>
