+++ src/components/grid/style.css
@@ -1,0 +1,10 @@
+.grid {
+  margin: 0 auto;
+}
+.gridItem {
+  margin-bottom: 14px;
+  width: 375px;
+}
+.gridSizer {
+  width: 375px;
+}
