+++ src/index.html
@@ -31,5 +31,5 @@
   <body>
     <div id="container"></div>
   </body>
-  <script src="app.js"></script>
+  <script src="index.js"></script>
 </html>
