+++ src/index.js
@@ -3,6 +3,8 @@ import ReactDOM from 'react-dom';
 import configureStore from './store/configure-store.js';
 import Root from './containers/root/index.js';
 
+require('./normalise.css');
+
 const store = configureStore();
 const rootElement = document.getElementById('container');
