+++ webpack.config.js
@@ -4,11 +4,11 @@ var webpack = require('webpack');
 
 module.exports = {
   entry: {
-    javascript: './src/app.js',
+    javascript: './src/index.js',
     html: './src/index.html'
   },
   output: {
-    filename: 'app.js',
+    filename: 'index.js',
     path: __dirname + '/dist'
   },
   resolve: {
