+++ webpack.production.config.js
@@ -6,7 +6,7 @@ var path = require('path');
 
 module.exports = {
   entry: [
-    './src/app.js'
+    './src/index.js'
   ],
   output: {
     path: __dirname + '/public/',
